--- api/app/lib/ai/ai.ts
@@ -1,3 +1,0 @@
-import OpenAI from 'openai'; // Uses process.env.OPENAI_API_KEY
-
-export const openai = new OpenAI();
