+++ api/app/lib/ai/index.ts
@@ -1,3 +1,2 @@
-export * from './ai';
 export * from './utils';
 export * from './queries';
